@@ -37,11 +37,11 @@
     "eslint": "^4.7.1",
     "eslint-loader": "^1.9.0",
     "eslint-plugin-react": "^7.3.0",
-    "mocha": "3.5.0",
-    "nodemon": "1.11.0",
+    "mocha": "3.5.3",
+    "nodemon": "1.12.1",
     "npm-run-all": "4.1.1",
     "style-loader": "0.18.2",
-    "webpack-dev-server": "2.7.1"
+    "webpack-dev-server": "2.9.1"
   },
   "dependencies": {
     "archiver": "2.0.3",
@@ -53,30 +53,30 @@
     "babel-preset-stage-2": "6.24.1",
     "classnames": "2.2.5",
     "ejs": "2.5.7",
-    "express": "4.15.4",
+    "express": "4.16.1",
     "extract-text-webpack-plugin": "3.0.0",
     "lodash": "4.17.4",
     "marked": "0.3.6",
     "picturefill": "3.0.2",
     "postcss-calc": "6.0.0",
-    "postcss-cli": "4.1.0",
+    "postcss-cli": "4.1.1",
     "postcss-custom-properties": "6.1.0",
-    "postcss-easy-import": "2.1.0",
+    "postcss-easy-import": "3.0.0",
     "postcss-functions": "3.0.0",
     "postcss-loader": "2.0.6",
     "postcss-nested": "2.1.2",
     "pretty": "2.0.0",
-    "prop-types": "15.5.10",
+    "prop-types": "15.6.0",
     "proxy-middleware": "0.15.0",
-    "react": "15.6.1",
-    "react-docgen": "2.17.0",
-    "react-dom": "15.6.1",
+    "react": "16.0.0",
+    "react-docgen": "2.18.0",
+    "react-dom": "16.0.0",
     "react-hot-loader": "3.0.0-beta.7",
     "react-redux": "5.0.6",
     "react-router-dom": "4.2.2",
     "react-simpletabs": "0.7.0",
     "redux": "3.7.2",
     "svgo": "0.7.2",
-    "webpack": "3.5.5"
+    "webpack": "3.6.0"
   }
 }
